Move glkernel stages to subfolder
