+++ source/gloperate-glkernel/include/gloperate-glkernel/stages/HemisphereDistributionKernelStage.h
