+++ source/gloperate-glkernel/include/gloperate-glkernel/stages/NoiseKernelStage.h
