+++ source/gloperate-glkernel/include/gloperate-glkernel/stages/TransparencyKernelStage.h
